Update app.py
By defaut flask will bind to localhost a.k.a. 127.0.0.1 which is only accessible to process running on the same machine. Docker containers are virtually different machines so only a process running within the same container would be able to reach it.

What you want to do is to tell flask to be reachable using the public ip of the container (public to the docker network). Since you don’t know which ip address your container will be assigned within the docker subnetwork, 0.0.0.0 says to flask to accept incoming connections no matter the network adapter or the ip address.